site-legal/requirements.txt
Django>=5.0.7, <5.1.0
psycopg2-binary == 2.9.9
Pillow>=10.4.0, <10.5.0
django-summernote==0.8.20.0